deepGLMfit {deepglm}	R Documentation

Train a deepGLM model

Usage

deepGLMfit(X,y, Lrate=0.01, Network=c(10,10) , BatchSize=5000,
           S=10, LRateFactor=10000, Momentum=0.6, Patience=100,
           MaxEpoch = 100, Verbose=10, Distribution='normal',
           WindowSize=100, Seed=NaN, Intercept=TRUE)

Arguments

X
Predictor variables X, specified as an n-by-p matrix, where n is the number of observations and p is the number of predictor variables. Each column of X represents one variable, and each row represents one observation.

Data type: matrix | numerical

y
Response variable y, specified as an n-by-1 vector, where n is the number of observations. Each entry in y is the response for the corresponding row of X.

Data type: vector | numerical

Lrate
The fix learning rate that is used for training. If the learning rate is too small, training will take a long time, but if it is too high, the training is likely to get stuck at a suboptimal result

Data type: scalar | numerical

Network
Neuron Network structure for deepGLM. In the current version, deepglm supports only 1 node for the output layer, users just need to provide a structure for hidden layers in a vector where each element in the vector is the number of nodes in the corresponding hidden layer. For example: Network=c(10,10,10) assigns deepGLM use 3 hidden layers with 10 nodes for each.

Data type: vector | positive integer

BatchSize
The size of the mini-batch used for each training iteration. For deepGLM, batch size should be a large number (e.g. 5000, 10000) compared to batch size in deep learning literature (e.g. 128, 256) Must be a positive integer equal or smaller than number of observations of training data.

Data type single | positive integer

S
The number of samples needed for Monte Carlo approximation of gradient of lower bound. Must be an positive integer

Data type: scalar | positive integer

LRateFactor
Down-scaling factor that is applied to the learning rate every time a certain number of iterations has passed. Must be a positive integer. For example, LrateFactor=100 means that after the first 100 iterations,learning rate will be multiplied with 100/t in each iteration, where t is the current number of iterations

Momentum
Momentum weight for stochastic gradient ascend. The momentum determines the contribution of the gradient step from the previous iteration to the current iteration of training. It must be a value between 0 and 1, where 0 will give no contribution from the previous step, and 1 will give a maximal contribution from the previous step. Must be between 0 and 1

Data type: scalar | numeric (0 to 1)

Patience
Number of consecutive times that the validation loss is allowed to be larger than or equal to the previously smallest loss before network training is stopped, used as an early stopping criterion. Must be a positive integer.

Data type: scalar | positive integer

MaxEpoch
The maximum number of epochs that will be used for training. An epoch is defined as the number of iterations needed for optimization algorithm to scan entire training data. Must be a positive integer.

Data type: scalar | positive integer

Verbose
Number of iterations that information on training progress will be printed to the command window each time. Set to 0 to disable this options.

Data type: scalar | integer

Distribution
Name of the distribution of the response, chosen from the following:
'normal' - Normal distribution (for continuous response)
'binomial' - Binomial distribution (for binary response)
'poisson' - Poisson distribution (for counting response)

Data type: string

WindowSize
Size of moving average window that used to smooth the VB lowerbound. Must be an positive integer

Data type: scalar | positive integer

Seed
Seeds the random number generator using the nonnegative integer. Must be a nonnegative integer.

Data type: scalar | nonnegative integer

Intercept
Set true (default) to add a column of 1 to predictor observation X matrix (play the role as intercept). If the data have already included the first '1' column, set 'Intercept' to false.

Data type: logical

Value

deepGLMfit returns a list including all input setting by users and output training information with a combination of following members:

out.weights
Estimated Mean of weights of Deep Neuron Network from input layer to the last hidden layer. This will be used to quickly doing point estimation for new data

out.beta
Estimated Mean of weights from the last hidden layer to output layer. This will be used to quickly doing point estimation for new data.

out.shrinkage
Matrix storing estimated values of group Lasso coefficients during training phase, used for variable selection.

out.lb
A vector storing values of VB lowerbound calculated in each iteration. In the training algorithm, the convergence of lowerbound is used as early stopping rule.

out.lbBar
A vector storing values of VB lowerbound after smoothed by a moving average window

out.vbMU
Estimated Mean of Variational Distribution. Used to do predictio interval estimation for new data

out.vbSIGMA
Estimated Covariance matrix of Variational Distribution. Used to do prediction interval estimation for new data.

out.b
Estimated factor loading vector. Used to calculate the estimated Covariance matrix of Variational Distribution vbSIGMA

out.c
Estimated vector of idiosyncratic noise standard deviation. Used to calculate the estimated Covariance matrix of Variational Distribution vbSIGMA

out.sigma2Alpha
Estimated shape parameter of Invert Gamma Variation Approximation for variance of noise

out.sigma2Beta
Estimated scale parameter of Invert Gamma Variation Approximation for variance of noise

out.sigma2Mean
Estimated mean of covariance of noise in each iteration during training phase

out.nparams
Total number of parameters in deepGLM have to be trained.

out.indexTrack
A vector to keep track the indexes of all weights in Deep Neuron Network. Used to reconstruct Deep Neuron Network from vbMU when doing prediction on new data

out.CPU
Total amount of training time (in second)

Author(s)

Nghia Nguyen: nghia.nguyen@sydney.edu.au
Minh-Ngoc Tran: minh-ngoc.tran@sydney.edu.au

References

Minh-Ngoc Tran,Nghia Nguyen, David J. Nott and Robert Kohn (2018) Bayesian Deep Net GLM and GLMM. arXiv:1805.10157 https://arxiv.org/abs/1805.10157

See Also

deepGLMpredict

Examples

## Train a deepGLM model on a simulation data (with continuous outcome)
## Create a simulation data with continuous outcomes
sigma <- 1              # Variance of noise
n <- 10000              # Number of observations
N <- 10                 # Number of covariates
X <- matrix(runif(n*N),n,N)     # Design matrix
# Use only 5 covariates to generate data
y <- 5 + 3*(X[,1] + 2*X[,2])^2 + 5*X[,3]*X[,4] + 2*X[,5] + sigma*runif(n)
ntest <- 2000           # Number of observation used for testing
Xtest <- X[1:ntest,]    # Extract data for test
ytest <- y[1:ntest]     # True value of test observations
X <- X[(ntest+1):n,]    # Observation for training
y <- y[(ntest+1):n]     # True values of training observations

## Fit simulation training data with deepGLMfit
deepGLMout <-deepGLMfit(X, y, Network=c(5,5), Seed=100, Verbose=1, MaxEpoch=2000)

## Make point prediction on test data using deepGLMpredict
print('----------------Prediction---------------')
Pred <- deepGLMpredict(deepGLMout,Xtest,y=ytest)
cat('PPS on test set using deepGLM is: ',Pred$pps,'\n')
cat('MSE on test set using deepGLM is: ',Pred$mse,'\n')

## Make prediction on test data with prediction interval
Pred1 <- deepGLMpredict(deepGLMout,Xtest,y=ytest,Interval=1,Nsample=1000)

[Package deepglm version 0.0.0.9000 Index]