deepGLMpredict {deepglm}	R Documentation

Predict response of deepGLM model

Usage

deepGLMpredict(mdl, X, y=NULL, Interval=0, Nsample=1000, Intercept=TRUE)

Arguments

mdl
Trained deepGLM model which is the output from deepGLMfit function.

X
Predictor variables X, specified as an n-by-p matrix, where n is the number of observations and p is the number of predictor variables. Each column of X represents one variable, and each row represents one observation.

Data type: matrix | numerical

y
Specify column of test responses. If this option is specified with true response column of new observations, deepGLMpredict will return prediction scores (PPS, MSE or Classification Rate) using true response column vector y

Data type: vector | numerical

Interval
Return prediction interval estimation for observations in test data X. By default, this predictive interval capability is disable (Interval=0). For example, Interval=1 will return one standard deviation predictive interval for new observations.

Data type: scalar | positive numerical

Nsample
Number of samples generated from posterior distribution of model parameters used to make prediction interval estimation for test data. Must be a positive integer

Data type: scalar | positive integer

Intercept
Set TRUE (default) to add a column of 1 to predictor observation X matrix (play the role as intercept). If the data have already included the first '1' column, set Intercept=FALSE

Data type: logical

Value

deepGLMpredict returns a list of prediction results on un-seen data with the following members:

yhat
Vector of point estimation for X. The length of yhat is equal to the number of observations of X.

yNN
Vector of output values when pass X through trained Deep Neuron Network. The length of yNN is the number of observation in X. If responses are normal, yNN is equal to yhat

yProb
Vector of output values when pass X through trained Deep Neuron Network. The length of yNN is the number of observation in X. If responses are normal, yNN is equal to yhat

pps
Partial Predictive Score loss of trained deepGLM on X. This is only available when ytest is specified.

mse
Mean Square Error of trained deepGLM on X. This is only available for Normal or Poisson responses and ytest is specified

accuracy
Classification rate of trained deepGLM classifier on X. This is only available for binary responses and ytest is specified.

interval
Prediction interval estimation of trained deepGLM on Xtest. interval is a two columns matrix where each row is an prediction interval of the corresponding observation of X. This is only available when 'Interval' option is specified with a non-zero value.

Author(s)

Nghia Nguyen: nghia.nguyen@sydney.edu.au
Minh-Ngoc Tran: minh-ngoc.tran@sydney.edu.au

References

Minh-Ngoc Tran,Nghia Nguyen, David J. Nott and Robert Kohn (2018) Bayesian Deep Net GLM and GLMM. arXiv:1805.10157 https://arxiv.org/abs/1805.10157

See Also

deepGLMfit

Examples

## Train and make prediction with deepGLM model on a simulation data (with continuous outcome)
## Create a simulation data with continuous outcomes
sigma <- 1              # Variance of noise
n <- 10000              # Number of observations
N <- 10                 # Number of covariates
X <- matrix(runif(n*N),n,N)     # Design matrix
# Use only 5 covariates to generate data
y <- 5 + 3*(X[,1] + 2*X[,2])^2 + 5*X[,3]*X[,4] + 2*X[,5] + sigma*runif(n)
ntest <- 2000           # Number of observation used for testing
Xtest <- X[1:ntest,]    # Extract data for test
ytest <- y[1:ntest]     # True value of test observations
X <- X[(ntest+1):n,]    # Observation for training
y <- y[(ntest+1):n]     # True values of training observations

## Fit simulation training data with deepGLMfit
deepGLMout <-deepGLMfit(X, y, Network=c(5,5), Seed=100, Verbose=1, MaxEpoch=2000)

## Make point prediction on test data using deepGLMpredict
print('----------------Prediction---------------')
Pred <- deepGLMpredict(deepGLMout,Xtest,y=ytest)
cat('PPS on test set using deepGLM is: ',Pred$pps,'\n')
cat('MSE on test set using deepGLM is: ',Pred$mse,'\n')

## Make prediction on test data with prediction interval
Pred1 <- deepGLMpredict(deepGLMout,Xtest,y=ytest,Interval=1,Nsample=1000)

[Package deepglm version 0.0.0.9000 Index]